backend/config.py
from dotenv import load_dotenv
load_dotenv()

import os
from pydantic_settings import BaseSettings

class Settings(BaseSettings):
    GOOGLE_API_KEY: str = os.getenv("GOOGLE_API_KEY", "")
    SUPABASE_URL: str = os.getenv("SUPABASE_URL", "")
    SUPABASE_KEY: str = os.getenv("SUPABASE_KEY", "")
    SUPABASE_JWT_SECRET: str = os.getenv("SUPABASE_JWT_SECRET", "")
    CHROMA_DB_PATH: str = "./chroma_db"
    
    class Config:
        env_file = ".env"

settings = Settings()
